// frontend/src/api-generated/models/taiSanPhanCungResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface TaiSanPhanCungResponse {
  id?: number;
  idDanhMucTaiSan?: number;
  tenDanhMucTaiSan?: string;
  idLoaiTaiSan?: number;
  tenLoaiTaiSan?: string;
  idHangSanXuat?: number;
  tenHangSanXuat?: string;
  maMau?: string;
  tenMau?: string;
  hinhAnh?: string;
  coTheThaoLap?: boolean;
  moTa?: string;
  trangThai?: string;
  thoiGianTao?: string;
  thoiGianCapNhat?: string;
}