// frontend/src/api-generated/models/taiSanPhanCungRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface TaiSanPhanCungRequest {
  idDanhMucTaiSan: number;
  idLoaiTaiSan: number;
  idHangSanXuat: number;
  /**
     * @minLength 0
     * @maxLength 50
     */
  maMau?: string;
  /**
     * @minLength 0
     * @maxLength 150
     */
  tenMau: string;
  /**
     * @minLength 0
     * @maxLength 255
     */
  hinhAnh?: string;
  coTheThaoLap: boolean;
  moTa?: string;
  /**
     * @minLength 0
     * @maxLength 30
     */
  trangThai: string;
}